* search:
  Loading blog, search, individual, and archive banners based on page.
  Adding archive header and styling

+++ wp-content/themes/rr/header.php
@@ -333,15 +333,28 @@ $(".nav > ul > li > .sub-menu").removeClass("visible-sm");
 
 
 
-			<?php if( is_single()) { 
+			<?php if( is_single() || is_archive() || is_page( 'rr-blog' ) ) { 
 			  //echo do_shortcode('[add_header_image title="Blog" image_url="/wp-content/uploads/2013/09/blog-bg.png"]'); 
 			  $blog_page_id = get_ID_by_slug('blog/rr-blog');
         $page_rr_blog = get_page($blog_page_id);
         $content = $page_rr_blog->post_content;
-        echo apply_filters('the_content', $page_rr_blog->post_content); 
+        $content = apply_filters('the_content', $page_rr_blog->post_content); 
+        //echo apply_filters('the_content', $page_rr_blog->post_content); 
+        
+        //if ( is_search() ) {
+        if ( isset($_GET['search']) ) {
+          $content = str_replace('Blog', 'Search', $content);
+          echo $content;
+        } else if ( is_archive() ) {
+          $content = str_replace('Blog', 'Archive', $content);
+          echo $content;
+        } else {
+          echo $content;
+        }
+        
 			}?>
 			
-      <?php if ( !is_single() && !get_post_meta($post->ID, 'banner', true) && function_exists('easingsliderpro') ) { 
+      <?php if ( !is_single() && !is_archive() && !get_post_meta($post->ID, 'banner', true) && function_exists('easingsliderpro') ) { 
           echo '<div class="slider home">';
           easingsliderpro( 1 ); 
           echo '</div>';
